refactor: DispatcherServlet 이 생성자로 필드를 초기화하도록 변경
